refactor: extract interface - MarketStatusChecker
