+++ src/main/java/co/fineants/api/domain/holding/service/streamer/ClosedMarketPortfolioStreamer.java
@@ -22,6 +22,6 @@ public class ClosedMarketPortfolioStreamer implements PortfolioStreamer {
 
 	@Override
 	public boolean supports(LocalDateTime time) {
-		return !stockMarketChecker.isMarketOpen(time);
+		return !stockMarketChecker.isOpen(time);
 	}
 }
