+++ src/test/java/co/fineants/api/domain/holding/domain/factory/MarketStatusBasedPortfolioStreamerFactoryTest.java
@@ -12,6 +12,7 @@ import org.mockito.Mockito;
 
 import co.fineants.api.domain.holding.service.MarketStatusChecker;
 import co.fineants.api.domain.holding.service.PortfolioHoldingService;
+import co.fineants.api.domain.holding.service.WeekdayMarketStatusChecker;
 import co.fineants.api.domain.holding.service.streamer.FluxIntervalPortfolioStreamer;
 import co.fineants.api.domain.holding.service.streamer.PortfolioStreamer;
 import co.fineants.api.global.common.time.LocalDateTimeService;
@@ -24,7 +25,7 @@ class MarketStatusBasedPortfolioStreamerFactoryTest {
 	@BeforeEach
 	void setUp() {
 		PortfolioHoldingService portfolioHoldingService = Mockito.mock(PortfolioHoldingService.class);
-		MarketStatusChecker stockMarketChecker = new MarketStatusChecker();
+		MarketStatusChecker stockMarketChecker = new WeekdayMarketStatusChecker();
 		int intervalSecond = 5;
 		int maxCount = 6;
 		List<PortfolioStreamer> streamers = List.of(
